Harden algolia initialization on active models
Related to 9c4e0bd3

AlgoliaSearch has to be included on inherited callback to proper
initialize algolia search (otherwise it override previous config on
previous models)
