+++ app/models/documentation_entry.rb
@@ -1,9 +1,6 @@
 # frozen_string_literal: true
 
-class DocumentationEntry
-  include ActiveModel::Model
-  include ActiveModelAlgoliaSearchable
-
+class DocumentationEntry < ApplicationAlgoliaSearchableActiveModel
   attr_accessor :section, :title, :content, :anchor
 
   def self.developers
